+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### ✨ Enhancements
 
 - Integrated Watchdog to enable automatic re-checking of Python files when changes are detected.
+- Added `autoformat-options` configuration option to let users specify command-line arguments to the Black formatting tool
 
 ### 💫 New checkers